refactor: request to auth
