+++ src/app/(auth)/signin/page.tsx
@@ -5,40 +5,29 @@
 'use client';
 
 import React from 'react';
-import { useRouter } from 'next/navigation';
 
 import { useForm } from 'react-hook-form';
-import { useMutation } from '@tanstack/react-query';
 import { ZodError, z } from 'zod';
-import { toast } from 'react-toastify';
 import Link from 'next/link';
-import { useCreateAccount } from '../../../hooks/auth';
-import { TypeAccount } from '../../../hooks/auth/useAuth/types';
+import { SigingProps } from '@/service/auth/siging';
+import { useSignin } from '../../../hooks/auth';
 import { Input, InputPassword } from '../../../components';
-
-type FormData = {
- email: string;
- name: string;
- password: number;
- type_account: string;
-};
 
 const schema = z.object({
   email: z.string().email('Email invalido').nonempty(),
   name: z.string(),
   password: z.string().min(6).nonempty(),
-  type_account: z.string().nonempty(),
+  typeAccount: z.string().nonempty(),
 });
 
 export default function Signin() {
-  const router = useRouter();
-  const { createAccountUser, createDocumentForUser } = useCreateAccount();
+  const { isLoading, createAccountUser } = useSignin();
 
   const {
     register,
     handleSubmit,
     formState: { errors },
-  } = useForm<FormData>({
+  } = useForm<SigingProps>({
     resolver: async (data) => {
       try {
         schema.parse(data);
@@ -52,34 +41,9 @@ export default function Signin() {
     },
   });
 
-  const { mutate, isLoading } = useMutation(createAccountUser, {
-    onSuccess: async (user) => {
-      await createDocumentForUser({
-        id: user.uid,
-        expirationTimeToken: (await user.getIdTokenResult()).expirationTime,
-        token: (await user.getIdTokenResult()).token,
-        email: user.email || '',
-        name: user.displayName || '',
-        typeAccount: user.photoURL as TypeAccount,
-      });
-      router.push(`/control/${user.uid}`);
-    },
-    onError: ({ message }: { message: string }) => {
-      if (message === 'Firebase: Error (auth/email-already-in-use).') {
-        toast.error('Este email já esta em uso', {
-          position: toast.POSITION.TOP_RIGHT,
-        });
-        return;
-      }
-      toast.error('Erro no Servidor. Tente mais tarde!', {
-        position: toast.POSITION.TOP_RIGHT,
-      });
-    },
-  });
-
   return (
     <div className="flex h-[100vh] justify-center items-center ">
-      <form onSubmit={handleSubmit((values: any) => mutate(values))}>
+      <form onSubmit={handleSubmit((values: SigingProps) => createAccountUser(values))}>
         <div className="flex flex-col gap-8 w-[400px] bg-gray-800 p-7 rounded-lg ">
           <Input
             label="Nome"
@@ -138,11 +102,11 @@ export default function Signin() {
                 <li className="w-full border-b sm:border-b-0 sm:border-r dark:border-gray-600" key={currency}>
                   <div className="flex items-center pl-3">
                     <input
-                      {...register('type_account', { required: true })}
+                      {...register('typeAccount', { required: true })}
                       id={`horizontal-list-radio-${currency.toLowerCase()}`}
                       type="radio"
                       value={currency}
-                      name="type_account"
+                      name="typeAccount"
                       className={`${currency === 'hybrid' ? 'cursor-not-allowed' : 'cursor-pointer'}  w-4 h-4 focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-700 dark:focus:ring-offset-gray-700 focus:ring-2 dark:bg-gray-600 dark:border-gray-500`}
                     />
                     <label htmlFor={`horizontal-list-radio-${currency.toLowerCase()}`} className="w-full py-3 ml-2 text-sm font-medium dark:text-gray-300">
