+++ src/hooks/auth/useAuth/types.ts
@@ -1,5 +1,3 @@
-'use client';
-
 export type TypeAccount = 'real' | 'euro' | 'hybrid'
 
 export interface UserData {
