+++ src/hooks/auth/useAuth/useForgetPassword.ts
@@ -1,23 +1,36 @@
-/* eslint-disable max-len */
-/* eslint-disable no-useless-catch */
-/* eslint-disable camelcase */
-
-'use client';
-
-import { sendPasswordResetEmail } from '@firebase/auth';
-import { auth } from '../../../service/firebase';
+import { useMutation } from '@tanstack/react-query';
+import { forgetPassword } from '@/service/auth/forgetPassword';
+import { useRouter } from 'next/navigation';
+import { toast } from 'react-toastify';
 
 const useForgetPassword = () => {
-  const forgetPassword = async ({ email }:{email: string}) => {
-    try {
-      await sendPasswordResetEmail(auth, email);
-    } catch (error) {
-      throw error;
-    }
-  };
+  const router = useRouter();
+  const { mutate: onForgetPassword, isLoading } = useMutation(
+    forgetPassword,
+    {
+      onSuccess: async () => {
+        toast.success('E-mail de recupeção enviado!', {
+          position: toast.POSITION.TOP_RIGHT,
+        });
+        router.push('/login');
+      },
+      onError: ({ message }: { message: string }) => {
+        if (message === 'Firebase: Error (auth/invalid-value-(email),-starting-an-object-on-a-scalar-field).') {
+          toast.error('E-mail inválido. Por favor, insira um endereço de e-mail válido.', {
+            position: toast.POSITION.TOP_RIGHT,
+          });
+        } else {
+          toast.error('Erro no Servidor. Tente mais tarde!', {
+            position: toast.POSITION.TOP_RIGHT,
+          });
+        }
+      },
+    },
+  );
 
   return {
-    forgetPassword,
+    onForgetPassword,
+    isLoading,
   };
 };
