+++ src/service/auth/logout.ts
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+import { signOut } from 'firebase/auth';
+import { auth } from '../firebase';
+
+export async function logout() {
+  await signOut(auth);
+}
